+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/fuji/structure/JobDescriptor.java
@@ -1,15 +1,28 @@
 package io.github.sakurawald.fuji.module.initializer.fuji.structure;
 
+import io.github.sakurawald.fuji.core.document.descriptor.interfaces.SourceModuleGetter;
+import io.github.sakurawald.fuji.core.manager.impl.module.ModuleManager;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import org.quartz.Job;
 import org.quartz.JobDetail;
+import org.quartz.JobKey;
 import org.quartz.Trigger;
 
 import java.util.List;
 
 @Data
 @AllArgsConstructor
-public class JobDescriptor {
+public class JobDescriptor implements SourceModuleGetter {
     public final JobDetail jobDetail;
     List<? extends Trigger> triggersOfJob;
+
+    @Override
+    public String getSourceModule() {
+        JobDetail jobDetail = this.jobDetail;
+        JobKey jobKey = jobDetail.getKey();
+        Class<? extends Job> jobClass = jobDetail.getJobClass();
+        return ModuleManager.computeModulePathAsString(jobClass.getName());
+    }
+
 }